+++ src/DSerfozo.RpcBindings/Contract/Analyze/IIdGenerator.cs
@@ -2,6 +2,6 @@
 {
     public interface IIdGenerator
     {
-        int GetNextId();
+        long GetNextId();
     }
 }
