+++ src/DSerfozo.RpcBindings/Contract/Marshaling/ICallbackFactory.cs
@@ -4,6 +4,6 @@ namespace DSerfozo.RpcBindings.Contract
 {
     public interface ICallbackFactory<TMarshal>
     {
-        object CreateCallback(int id, Type delegateType, IParameterBinder<TMarshal> parameterBinder);
+        object CreateCallback(long id, Type delegateType, IParameterBinder<TMarshal> parameterBinder);
     }
 }
