+++ src/DSerfozo.RpcBindings/Marshaling/Callback.cs
@@ -4,7 +4,7 @@ namespace DSerfozo.RpcBindings.Marshaling
 {
     public class Callback<TMarshal> : CallbackBase<TMarshal>, ICallback
     {
-        public Callback(int id, ICallbackExecutor<TMarshal> executor, IParameterBinder<TMarshal> parameterBinder) 
+        public Callback(long id, ICallbackExecutor<TMarshal> executor, IParameterBinder<TMarshal> parameterBinder) 
             : base(id, executor, parameterBinder)
         {
         }
