+++ src/DSerfozo.RpcBindings/Model/MethodDescriptor.cs
@@ -11,7 +11,7 @@ namespace DSerfozo.RpcBindings.Model
         {
             private readonly MethodDescriptor constructed;
 
-            public Builder WithId(int id)
+            public Builder WithId(long id)
             {
                 constructed.Id = id;
                 return this;
@@ -59,7 +59,7 @@ namespace DSerfozo.RpcBindings.Model
         }
 
         [ShouldSerialize]
-        public int Id { get; private set;  }
+        public long Id { get; private set;  }
 
         [ShouldSerialize]
         public string Name { get; private set; }
